[#17] Refactor: Fixtures util 패키지로 이동
