+++ src/test/groovy/com/devtribe/devtribe_feed_service/test/utils/fixtures/CreatePostRequestFixture.groovy
@@ -1,4 +1,4 @@
-package com.devtribe.devtribe_feed_service.test.fixtures
+package com.devtribe.devtribe_feed_service.test.utils.fixtures
 
 import com.devtribe.devtribe_feed_service.post.application.dtos.CreatePostRequest
